Feature: Landing page for Shop
Why?:
Homepage for guests/users visiting the shop site.

This addresses the need by:
Designing views and adding a controller to render the landing page of the shop website.

Header and Footer are seperated into partials for easy use with other pages.

Adding products and categories to be shown on homepage as an array from the controller.

Pages were designed with the Materialize CSS Framework.

[Finishes #108000292]