Designed the 'Product Description' page.
Why?
When a product is click, we need a page where all the details of the
specific product is displayed. From this page, shop cusotomers can decide
to buy the product, add it to their wishlist and see other related
products.

How?
Create a Products controller,
Create the 'show product' view,
Create the route that points to this page
Create the scss stiesheet file for this view
Create a js file that initializes some scripts is this page.

Completed 'Design Product Description Page':
https://www.pivotaltracker.com/story/show/108000136
finished [#108000136]