Fix Signin and Signup links
Why?
When the signin link on the landing page is clicked or the signup in the
signin page, the signin/signup page should be loaded.

How?
Create route to the signin page
Use link_to to add the link to the signin page
Create route to the signup page,
Add signup link to the signin page
Add signin link to the signup page.