Create Customer Account
Why?
When a user fills the sign up form
His/her data should be stored in the database
Add a welcome message containing the activation link shoulld be sent
to the user's email address

How?
Generate user migration, model and controller
Write the script to process the submitted form and store the valid data to the database
Generate action mailer to send the user a welcome mail with the activation link.

https://www.pivotaltracker.com/story/show/110060500
[finished #110060500]