Add to Wishlist Functionality
Why?
Users should be able to add or save items to their wishlist for easy retrieval whenever they're ready to purchase the item.

How?
Created the Wishlist model, which references the user, and the product. Each new entry is stored with a reference to the user it belongs to, and the product it saves.
Set up Wishlist controller which handles managing wishlists.

https://www.pivotaltracker.com/story/show/110082496