Implement Account Activation
Why?
When a user creates an account,
An activation link is sent to the user's email address.
When the user clicks on this activation link
His/Her account should be activated
The user also gets a confirmation flash that the account has been activated

How?
Impement the `activate` action in `UsersController`
Search the user with the given `id` and `activation_token`
Update the user's `activate_status`

Completing https://www.pivotaltracker.com/story/show/112346625
[finished #112346625]