Display Related Products
Why?
GIVEN that I am a customer
AND I'm in the the product description page
I SHOULD be able to see other products that are related to the current product

How?
implement product controller to select and display products in
the same category as that of the current product

https://www.pivotaltracker.com/story/show/110084854

[finishes #110084854]

displays related products