Display Product Reviews and Ratings [Finishes #110250232]
Why?
Customers viewing a product's details should be able to view reviews or ratings given by other customers. This gives them first-hand information about the product from people who might have bought it, or used it.

How?
Added a Review model which stores a comment/rating pair for a particular product. It's linked to a product, and the user who made the comment.

https://www.pivotaltracker.com/story/show/110250232