Update Cart from the cart page.
Why?
When a customer is done add items to cart, and proceed to cart
He/she should be able to remove items from the cart and continue the checkout process with the item and quantity in the cart

How?
Use cookies hash to store the items added to the cart so as to build the cart order items form from it.

https://www.pivotaltracker.com/story/show/112511965